+++ secrets/consul/backend_test.go
@@ -355,6 +355,52 @@ func testBackendRenewRevoke14(t *testing.T, version string, policiesParam string
 	if err == nil {
 		t.Fatal("err: expected error")
 	}
+
+	t.Run("revoking missing token", func(t *testing.T) {
+		// read new token
+		req.Operation = logical.ReadOperation
+		req.Path = "creds/test"
+		resp, err = b.HandleRequest(context.Background(), req)
+		if err != nil {
+			t.Fatal(err)
+		}
+		if resp == nil {
+			t.Fatal("resp nil")
+		}
+		if resp.IsError() {
+			t.Fatalf("resp is error: %v", resp.Error())
+		}
+
+		if err := mapstructure.Decode(resp.Data, &d); err != nil {
+			t.Fatal(err)
+		}
+
+		// Delete token using consul api
+		consulapiConfig.Token = consulConfig.Token
+		client, err = consulapi.NewClient(consulapiConfig)
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		_, err = client.ACL().TokenDelete(d.Accessor, &consulapi.WriteOptions{})
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		// revoke
+		req.Operation = logical.RevokeOperation
+		req.Secret = resp.Secret
+		_, err = b.HandleRequest(context.Background(), req)
+		if err != nil {
+			t.Fatal(err)
+		}
+		if resp == nil {
+			t.Fatal("resp nil")
+		}
+		if resp.IsError() {
+			t.Fatalf("resp is error: %v", resp.Error())
+		}
+	})
 }
 
 func TestBackend_LocalToken(t *testing.T) {